display_task: Add shared dispatch module
Extract the FreeRTOS queue, consume_mailbox callback, and
process_messages task into a shared display_task module.  All
seven ESP32 drivers now embed struct DisplayTaskArgs and use
the shared dispatch infrastructure via CONTAINER_OF.

The consume_mailbox implementation adopts the ili948x variant
(non-blocking enqueue with drop-oldest-on-overflow and proper
disposal of dropped messages).

Also removes unused display_enqueue_message from ili934x and
stale mailbox.h include from display_driver.h.
