+++ ili948x_display_driver.c
@@ -55,6 +55,7 @@
 #include "display_common.h"
 #include "display_items.h"
 #include "display_message.h"
+#include "display_task.h"
 #include "image_helpers.h"
 #include "spi_display.h"
 
@@ -119,7 +120,12 @@ struct SPI
     bool madctl_bgr;
 
     Context *ctx;
+
+    struct DisplayTaskArgs display_args;
 };
+
+#define SPI_FROM_CTX(ctx) \
+    CONTAINER_OF((struct DisplayTaskArgs *) (ctx)->platform_data, struct SPI, display_args)
 
 // Double-buffered scanline buffers.
 struct Screen
@@ -196,9 +202,6 @@ static inline void rgb565swapped_line_to_rgb888(uint8_t *dst, const uint16_t *sr
     }
 }
 
-static QueueHandle_t display_messages_queue;
-
-static NativeHandlerResult display_driver_consume_mailbox(Context *ctx);
 static void display_init(Context *ctx, term opts);
 
 static void display_init9486(struct SPI *spi);
@@ -485,7 +488,7 @@ static void do_update(Context *ctx, term display_list)
 
     int screen_width = screen->w;
     int screen_height = screen->h;
-    struct SPI *spi = ctx->platform_data;
+    struct SPI *spi = SPI_FROM_CTX(ctx);
 
     set_screen_paint_area(spi, 0, 0, screen_width, screen_height);
     writecommand(spi, ILI948X_RAMWR);
@@ -618,7 +621,7 @@ static void process_message(Message *message, Context *ctx)
     }
     term cmd = term_get_tuple_element(req, 0);
 
-    struct SPI *spi = ctx->platform_data;
+    struct SPI *spi = SPI_FROM_CTX(ctx);
 
     if (cmd == context_make_atom(ctx, "\x6"
                                       "update")) {
@@ -659,46 +662,6 @@ static void process_message(Message *message, Context *ctx)
 
     send_message(gen_message.pid, return_tuple, ctx->global);
     END_WITH_STACK_HEAP(heap, ctx->global);
-}
-
-static void process_messages(void *arg)
-{
-    struct SPI *args = arg;
-
-    while (true) {
-        Message *message;
-        xQueueReceive(display_messages_queue, &message, portMAX_DELAY);
-        process_message(message, args->ctx);
-
-        BEGIN_WITH_STACK_HEAP(1, temp_heap);
-        mailbox_message_dispose(&message->base, &temp_heap);
-        END_WITH_STACK_HEAP(temp_heap, args->ctx->global);
-    }
-}
-
-static NativeHandlerResult display_driver_consume_mailbox(Context *ctx)
-{
-    MailboxMessage *mbox_msg = mailbox_take_message(&ctx->mailbox);
-    Message *msg = CONTAINER_OF(mbox_msg, Message, base);
-
-    // Non-blocking enqueue; drop oldest on overflow.
-    if (xQueueSend(display_messages_queue, &msg, 0) != pdTRUE) {
-
-        Message *old = NULL;
-        if (xQueueReceive(display_messages_queue, &old, 0) == pdTRUE && old) {
-            BEGIN_WITH_STACK_HEAP(1, temp_heap);
-            mailbox_message_dispose(&old->base, &temp_heap);
-            END_WITH_STACK_HEAP(temp_heap, ctx->global);
-        }
-
-        if (xQueueSend(display_messages_queue, &msg, 0) != pdTRUE) {
-            BEGIN_WITH_STACK_HEAP(1, temp_heap2);
-            mailbox_message_dispose(&msg->base, &temp_heap2);
-            END_WITH_STACK_HEAP(temp_heap2, ctx->global);
-        }
-    }
-
-    return NativeContinue;
 }
 
 static void set_rotation(struct SPI *spi, int rotation)
@@ -743,10 +706,12 @@ static void display_init(Context *ctx, term opts)
 {
     screen = malloc(sizeof(struct Screen));
 
-    display_messages_queue = xQueueCreate(32, sizeof(Message *));
-
     struct SPI *spi = malloc(sizeof(struct SPI));
-    ctx->platform_data = spi;
+
+    spi->display_args.messages_queue = xQueueCreate(32, sizeof(Message *));
+    spi->display_args.process_message_fn = process_message;
+    spi->display_args.ctx = ctx;
+    ctx->platform_data = &spi->display_args;
 
     spi->ctx = ctx;
 
@@ -872,7 +837,7 @@ static void display_init(Context *ctx, term opts)
     backlight_gpio_parse_config(&backlight_config, opts, ctx->global);
     backlight_gpio_init(&backlight_config);
 
-    xTaskCreate(process_messages, "display", 10000, spi, 1, NULL);
+    xTaskCreate(display_process_messages, "display", 10000, &spi->display_args, 1, NULL);
 }
 
 static void display_init9486(struct SPI *spi)
